refactor: remove useCreateHeadingLink hook
- change to direct implementation in MarkdownView component
- change createRoot to renderToStaticMarkup
